Give zeus access to server perf stats (merge zeusMonitor into serverMonitor)
