+++ addons/serverMonitor/RscDisplayCurator.hpp
@@ -11,7 +11,7 @@ class RscDisplayCurator {
             h = "12 * pixelH";
             sizeEx = "12 * pixelH";
             text = "ACRE Spectator";
-            action = QUOTE(_this call FUNC(toggleAcreSepcator));
+            action = QUOTE(_this call FUNC(zeus_toggleAcreSepcator));
         };
         class GVAR(unitCount): RscText {
             idc = 777551;
