+++ addons/serverMonitor/XEH_PREP.hpp
@@ -1,0 +1,4 @@
+TRACE_1("",QUOTE(ADDON));
+
+PREP(zeus_toggleAcreSepcator);
+PREP(zeus_interfaceOpenedEH);
