+++ addons/settings/config.cpp
@@ -5,11 +5,7 @@ class CfgPatches {
         units[] = {};
         weapons[] = {};
         requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {
-            "potato_core","ace_laserPointer","ace_map","ace_zeus","ace_hearing",
-            "ace_overheating","ace_gforces","ace_goggles","ace_frag",
-            "ace_medical_ai","ace_advanced_fatigue", "ace_cookoff", "STUI_GroupHUD"
-        };
+        requiredAddons[] = {"potato_core"};
         author = "Potato";
         authors[] = {"PabstMirror", "AACO"};
         authorUrl = "https://github.com/BourbonWarfare/POTATO";
