--- addons/zeusMonitor/$PBOPREFIX$
@@ -1,1 +1,0 @@
-z\potato\addons\zeusMonitor
