--- addons/zeusMonitor/XEH_PREP.hpp
@@ -1,4 +1,0 @@
-TRACE_1("",QUOTE(ADDON));
-
-PREP(toggleAcreSepcator);
-PREP(zeusInterfaceOpenedEH);
