--- addons/zeusMonitor/config.cpp
@@ -1,17 +1,0 @@
-#include "script_component.hpp"
-
-class CfgPatches {
-    class ADDON {
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = {"potato_core"};
-        author = "Potato";
-        authors[] = {"PabstMirror"};
-        authorUrl = "https://github.com/BourbonWarfare/POTATO";
-        VERSION_CONFIG;
-    };
-};
-
-#include "CfgEventHandlers.hpp"
-#include "RscDisplayCurator.hpp"
