--- addons/zeusMonitor/functions/script_component.hpp
@@ -1,1 +1,0 @@
-#include "\z\potato\addons\zeusMonitor\script_component.hpp"
