--- addons/zeusMonitor/script_component.hpp
@@ -1,12 +1,0 @@
-#define COMPONENT zeusMonitor
-#include "\z\potato\addons\core\script_mod.hpp"
-
-// #define DEBUG_MODE_FULL
-// #define DISABLE_COMPILE_CACHE
-// #define ENABLE_PERFORMANCE_COUNTERS
-
-#ifdef DEBUG_ENABLED_ZEUSMONITOR
-    #define DEBUG_MODE_FULL
-#endif
-
-#include "\z\potato\addons\core\script_macros.hpp"
